requirements.txt
alembic==1.16.2
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
defusedxml==0.7.1
exceptiongroup==1.3.0
fastapi==0.115.14
fonttools==4.58.4
fpdf2==2.8.3
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
greenlet==3.2.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
six @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.35.0
